[feat] user info is displayed in watch mode
